Module arrow_string::like

Macros

dict_function 🔒

dyn_function 🔒

scalar_dyn_function 🔒

Functions

compare_op 🔒

Helper function to perform boolean lambda function on values from two array accessors, this version does not attempt to use SIMD.

compare_op_scalar 🔒

Helper function to perform boolean lambda function on values from array accessor, this version does not attempt to use SIMD.

contains 🔒

contains_dict 🔒

Perform SQL CONTAINS(left, right) operation on [DictionaryArray] with values [StringArray]/[LargeStringArray].

contains_dyn

Perform SQL CONTAINS(left, right) operation on [StringArray] / [LargeStringArray], or [DictionaryArray] with values [StringArray]/[LargeStringArray].

contains_scalar 🔒

contains_utf8

Perform SQL CONTAINS(left, right) operation on [StringArray] / [LargeStringArray].

contains_utf8_scalar

Perform SQL CONTAINS(left, right) operation on [StringArray] / [LargeStringArray] and a scalar.

contains_utf8_scalar_dyn

Perform SQL CONTAINS(left, right) operation on [StringArray] / [LargeStringArray], or [DictionaryArray] with values [StringArray]/[LargeStringArray] and a scalar.

ends_with 🔒

ends_with_dict 🔒

Perform SQL ENDSWITH(left, right) operation on [DictionaryArray] with values [StringArray]/[LargeStringArray].

ends_with_dyn

Perform SQL ENDSWITH(left, right) operation on [StringArray] / [LargeStringArray], or [DictionaryArray] with values [StringArray]/[LargeStringArray].

ends_with_scalar 🔒

ends_with_utf8

Perform SQL ENDSWITH(left, right) operation on [StringArray] / [LargeStringArray].

ends_with_utf8_scalar

Perform SQL ENDSWITH(left, right) operation on [StringArray] / [LargeStringArray] and a scalar.

ends_with_utf8_scalar_dyn

Perform SQL ENDSWITH(left, right) operation on [StringArray] / [LargeStringArray], or [DictionaryArray] with values [StringArray]/[LargeStringArray] and a scalar.

ilike 🔒

ilike_dict 🔒

Perform SQL left ILIKE right operation on [DictionaryArray] with values [StringArray]/[LargeStringArray].

ilike_dyn

Perform SQL left ILIKE right operation on [StringArray] / [LargeStringArray], or [DictionaryArray] with values [StringArray]/[LargeStringArray].

ilike_scalar 🔒

ilike_scalar_op 🔒

ilike_utf8

Perform SQL left ILIKE right operation on [StringArray] / [LargeStringArray].

ilike_utf8_scalar

Perform SQL left ILIKE right operation on [StringArray] / [LargeStringArray] and a scalar.

ilike_utf8_scalar_dyn

Perform SQL left ILIKE right operation on [StringArray] / [LargeStringArray], or [DictionaryArray] with values [StringArray]/[LargeStringArray] and a scalar.

is_like_pattern 🔒

like 🔒

like_dict 🔒

Perform SQL left LIKE right operation on [DictionaryArray] with values [StringArray]/[LargeStringArray].

like_dyn

Perform SQL left LIKE right operation on [StringArray] / [LargeStringArray], or [DictionaryArray] with values [StringArray]/[LargeStringArray].

like_scalar 🔒

like_scalar_op 🔒

like_utf8

Perform SQL left LIKE right operation on [StringArray] / [LargeStringArray].

like_utf8_scalar

Perform SQL left LIKE right operation on [StringArray] / [LargeStringArray] and a scalar.

like_utf8_scalar_dyn

Perform SQL left LIKE right operation on [StringArray] / [LargeStringArray], or [DictionaryArray] with values [StringArray]/[LargeStringArray] and a scalar.

nilike 🔒

nilike_dict 🔒

Perform SQL left NOT ILIKE right operation on [DictionaryArray] with values [StringArray]/[LargeStringArray].

nilike_dyn

Perform SQL left NOT ILIKE right operation on [StringArray] / [LargeStringArray], or [DictionaryArray] with values [StringArray]/[LargeStringArray].

nilike_scalar 🔒

nilike_utf8

Perform SQL left NOT ILIKE right operation on [StringArray] / [LargeStringArray].

nilike_utf8_scalar

Perform SQL left NOT ILIKE right operation on [StringArray] / [LargeStringArray] and a scalar.

nilike_utf8_scalar_dyn

Perform SQL left NOT ILIKE right operation on [StringArray] / [LargeStringArray], or [DictionaryArray] with values [StringArray]/[LargeStringArray] and a scalar.

nlike 🔒

nlike_dict 🔒

Perform SQL left NOT LIKE right operation on [DictionaryArray] with values [StringArray]/[LargeStringArray].

nlike_dyn

Perform SQL left NOT LIKE right operation on [StringArray] / [LargeStringArray], or [DictionaryArray] with values [StringArray]/[LargeStringArray].

nlike_scalar 🔒

nlike_utf8

Perform SQL left NOT LIKE right operation on [StringArray] / [LargeStringArray].

nlike_utf8_scalar

Perform SQL left NOT LIKE right operation on [StringArray] / [LargeStringArray] and a scalar.

nlike_utf8_scalar_dyn

Perform SQL left NOT LIKE right operation on [StringArray] / [LargeStringArray], or [DictionaryArray] with values [StringArray]/[LargeStringArray] and a scalar.

regex_like 🔒

Evaluate regex op(left) matching right on [StringArray] / [LargeStringArray]

replace_like_wildcards 🔒

Transforms a like pattern to a regex compatible pattern. To achieve that, it does:

starts_with 🔒

starts_with_dict 🔒

Perform SQL STARTSWITH(left, right) operation on [DictionaryArray] with values [StringArray]/[LargeStringArray].

starts_with_dyn

Perform SQL STARTSWITH(left, right) operation on [StringArray] / [LargeStringArray], or [DictionaryArray] with values [StringArray]/[LargeStringArray].

starts_with_scalar 🔒

starts_with_utf8

Perform SQL STARTSWITH(left, right) operation on [StringArray] / [LargeStringArray].

starts_with_utf8_scalar

Perform SQL STARTSWITH(left, right) operation on [StringArray] / [LargeStringArray] and a scalar.

starts_with_utf8_scalar_dyn

Perform SQL STARTSWITH(left, right) operation on [StringArray] / [LargeStringArray], or [DictionaryArray] with values [StringArray]/[LargeStringArray] and a scalar.