Trait arrow_string::like::StringArrayType

trait StringArrayType<'a>: ArrayAccessor<Item = &'a str> + Sized {
    // Required methods
    fn is_ascii(&self) -> bool;
    fn iter(&self) -> ArrayIter<Self>;
}

A trait for Arrow String Arrays, currently three types are supported:

StringArray

LargeStringArray

StringViewArray

This trait helps to abstract over the different types of string arrays so that we don’t need to duplicate the implementation for each type.

Required Methods

fn is_ascii(&self) -> bool

fn iter(&self) -> ArrayIter<Self>

Object Safety

This trait is not object safe.