arrow_string

Module regexp

Defines kernel to extract substrings based on a regular expression of a [Large]StringArray

Macros

process_regexp_array_match 🔒

process_regexp_match 🔒

Functions

get_scalar_pattern_flag 🔒

get_scalar_pattern_flag_utf8view 🔒

regexp_array_match 🔒

regexp_array_match_utf8view 🔒

regexp_is_match

Return BooleanArray indicating which strings in an array match an array of regular expressions.

regexp_is_match_scalar

Return BooleanArray indicating which strings in an array match a single regular expression.

regexp_is_match_utf8Deprecated

Perform SQL array ~ regex_array operation on [StringArray] / [LargeStringArray]. If regex_array element has an empty value, the corresponding result value is always true.

regexp_is_match_utf8_scalarDeprecated

Perform SQL array ~ regex_array operation on [StringArray] / [LargeStringArray] and a scalar.

regexp_match

Extract all groups matched by a regular expression for a given String array.

regexp_scalar_match 🔒

regexp_scalar_match_utf8view 🔒