Crate arrow_string

Arrow string kernels

Modules

concat_elements

Provides utility functions for concatenation of elements in arrays.

length

Defines kernel for length of string arrays and binary arrays

like

Provide SQL’s LIKE operators for Arrow’s string arrays

predicate 🔒

regexp

Defines kernel to extract substrings based on a regular expression of a [Large]StringArray

substring

Defines kernel to extract a substring of an Array Supported array types: [GenericStringArray], [GenericBinaryArray], [FixedSizeBinaryArray], [DictionaryArray]