Crate arrow_string

Arrow string kernels

Modules

concat_elements

length

Defines kernel for length of string arrays and binary arrays

like

regexp

Defines kernel to extract substrings based on a regular expression of a [Large]StringArray

substring

Defines kernel to extract a substring of an Array Supported array types: GenericStringArray, GenericBinaryArray, FixedSizeBinaryArray, DictionaryArray