arrow_string
Module length

Defines kernel for length of string arrays and binary arrays

Functions

bit_length

Returns an array of Int32/Int64 denoting the number of bits in each value in the array.

bit_length_impl 🔒

length

Returns an array of Int32/Int64 denoting the length of each value in the array.

length_impl 🔒